Move TypeScript source file directly in lib
