+++ .gitignore
@@ -1,18 +1,16 @@
 node_modules/
 .idea/
 *.iml
-lib/
 test/**/*.js
 test/**/*.map
 test/**/*.d.ts
-src/**/*.js
-src/**/*.map
-src/**/*.d.ts
+lib/**/*.js
+lib/**/*.map
+lib/**/*.d.ts
 coverage/
 .nyc_output
 etc
 priv
 .travis.yml
 tslint.json
-lib/**/*.map
 package-lock.json
