+++ lib/digest-auth.ts
