+++ lib/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "../tsconfig.json",
   "compilerOptions": {
     "sourceMap": true,
-    "outDir": "../lib",
     "declaration": true
   }
 }
