+++ lib/xml/xml-metadata.ts
